# Conflicts:
#   backend/src/main/java/ssafy/backend/afterglow/controller/RecordController.java
#   backend/src/main/java/ssafy/backend/afterglow/domain/ImageRecord.java
#   backend/src/main/java/ssafy/backend/afterglow/repository/RouteRepository.java
#   backend/src/main/java/ssafy/backend/afterglow/service/RecordService.java

+++ backend/src/main/java/ssafy/backend/afterglow/repository/ConsumptionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ssafy.backend.afterglow.domain.ConsumptionRecord;
 import ssafy.backend.afterglow.domain.Record;
 
-public interface ConsumptionRepository extends JpaRepository<ConsumptionRecord, Integer> {
+public interface ConsumptionRepository extends JpaRepository<ConsumptionRecord, Long> {
 }
